Refactor Beam and remove unused code. Upstream BTLx related properties to TimberElement
