+++ src/compas_timber/elements/timber.py
@@ -1,3 +1,5 @@
+from compas.geometry import Frame
+from compas.geometry import Line
 from compas.geometry import Transformation
 from compas_model.elements import Element
 from compas_model.elements import reset_computed
@@ -112,3 +114,123 @@ class TimberElement(Element):
                 features = [features]
             self._features = [f for f in self._features if f not in features]
         self._geometry = None  # reset geometry cache TODO: should we do that?
+
+    ########################################################################
+    # BTLx properties
+    ########################################################################
+
+    @property
+    def ref_frame(self):
+        # type: () -> Frame
+        """Reference frame for machining processings according to BTLx standard. The origin is at the center of the beam."""
+        assert self.frame
+        start, _ = self._resolve_blank_extensions()
+        ref_point = self.frame.point.copy()
+        ref_point += -self.frame.xaxis * start  # "extension" to the start edge
+
+        ref_point += self.frame.yaxis * self.width * 0.5
+        ref_point -= self.frame.zaxis * self.height * 0.5
+        return Frame(ref_point, self.frame.xaxis, self.frame.zaxis)
+
+    @property
+    def ref_sides(self):
+        # type: () -> tuple[Frame, Frame, Frame, Frame, Frame, Frame]
+        # See: https://design2machine.com/btlx/BTLx_2_2_0.pdf
+        # TODO: cache these
+        rs1_point = self.ref_frame.point
+        rs2_point = rs1_point + self.ref_frame.yaxis * self.height
+        rs3_point = rs1_point + self.ref_frame.yaxis * self.height + self.ref_frame.zaxis * self.width
+        rs4_point = rs1_point + self.ref_frame.zaxis * self.width
+        rs5_point = rs1_point
+        rs6_point = rs1_point + self.ref_frame.xaxis * self.blank_length + self.ref_frame.yaxis * self.height
+        return (
+            Frame(rs1_point, self.ref_frame.xaxis, self.ref_frame.zaxis, name="RS_1"),
+            Frame(rs2_point, self.ref_frame.xaxis, -self.ref_frame.yaxis, name="RS_2"),
+            Frame(rs3_point, self.ref_frame.xaxis, -self.ref_frame.zaxis, name="RS_3"),
+            Frame(rs4_point, self.ref_frame.xaxis, self.ref_frame.yaxis, name="RS_4"),
+            Frame(rs5_point, self.ref_frame.zaxis, self.ref_frame.yaxis, name="RS_5"),
+            Frame(rs6_point, self.ref_frame.zaxis, -self.ref_frame.yaxis, name="RS_6"),
+        )
+
+    @property
+    def ref_edges(self):
+        # type: () -> tuple[Line, Line, Line, Line]
+        # so tuple is not created every time
+        ref_sides = self.ref_sides
+        return (
+            Line(ref_sides[0].point, ref_sides[0].point + ref_sides[0].xaxis * self.blank_length, name="RE_1"),
+            Line(ref_sides[1].point, ref_sides[1].point + ref_sides[1].xaxis * self.blank_length, name="RE_2"),
+            Line(ref_sides[2].point, ref_sides[2].point + ref_sides[2].xaxis * self.blank_length, name="RE_3"),
+            Line(ref_sides[3].point, ref_sides[3].point + ref_sides[3].xaxis * self.blank_length, name="RE_4"),
+        )
+
+    def front_side(self, ref_side_index):
+        # type: (int) -> Frame
+        """Returns the next side after the reference side, following the right-hand rule with the thumb along the beam's frame x-axis.
+        This method does not consider the start and end sides of the beam (RS5 & RS6).
+
+        Parameters
+        ----------
+        ref_side_index : int
+            The index of the reference side to which the front side should be calculated.
+
+        Returns
+        -------
+        frame : :class:`~compas.geometry.Frame`
+            The frame of the front side of the beam relative to the reference side.
+        """
+        return self.ref_sides[(ref_side_index + 1) % 4]
+
+    def back_side(self, ref_side_index):
+        # type: (int) -> Frame
+        """Returns the previous side before the reference side, following the right-hand rule with the thumb along the beam's frame x-axis.
+        This method does not consider the start and end sides of the beam (RS5 & RS6).
+
+        Parameters
+        ----------
+        ref_side_index : int
+            The index of the reference side to which the back side should be calculated.
+
+        Returns
+        -------
+        frame : :class:`~compas.geometry.Frame`
+            The frame of the back side of the beam relative to the reference side.
+        """
+        return self.ref_sides[(ref_side_index - 1) % 4]
+
+    def opp_side(self, ref_side_index):
+        # type: (int) -> Frame
+        """Returns the the side that is directly across from the reference side, following the right-hand rule with the thumb along the beam's frame x-axis.
+        This method does not consider the start and end sides of the beam (RS5 & RS6).
+
+        Parameters
+        ----------
+        ref_side_index : int
+            The index of the reference side to which the opposite side should be calculated.
+
+        Returns
+        -------
+        frame : :class:`~compas.geometry.Frame`
+            The frame of the opposite side of the beam relative to the reference side.
+        """
+        return self.ref_sides[(ref_side_index + 2) % 4]
+
+    def get_dimensions_relative_to_side(self, ref_side_index):
+        # type: (int) -> tuple[float, float]
+        """Returns the perpendicular and parallel dimensions of the beam to the given reference side.
+
+        Parameters
+        ----------
+        ref_side_index : int
+            The index of the reference side to which the dimensions should be calculated.
+
+        Returns
+        -------
+        tuple(float, float)
+            The perpendicular and parallel dimensions of the beam to the reference side.
+                - Perpendicular dimension: The measurement normal to the reference side.
+                - Parallel dimension: The measurement along y-axis of reference side.
+        """
+        if ref_side_index in [1, 3]:
+            return self.height, self.width
+        return self.width, self.height
